New permission and role features

+++ auth-service/src/main/java/com/ttdat/authservice/application/eventhandlers/PermissionEventHandler.java
@@ -1,6 +1,7 @@
 package com.ttdat.authservice.application.eventhandlers;
 
 import com.ttdat.authservice.application.exception.ErrorCode;
+import com.ttdat.authservice.application.exception.ResourceInUseException;
 import com.ttdat.authservice.application.exception.ResourceNotFoundException;
 import com.ttdat.authservice.application.mappers.PermissionMapper;
 import com.ttdat.authservice.domain.entities.Permission;
@@ -12,6 +13,7 @@ import lombok.RequiredArgsConstructor;
 import org.axonframework.config.ProcessingGroup;
 import org.axonframework.eventhandling.EventHandler;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 @Component
 @RequiredArgsConstructor
@@ -20,12 +22,14 @@ public class PermissionEventHandler {
     private final PermissionRepository permissionRepository;
     private final PermissionMapper permissionMapper;
 
+    @Transactional
     @EventHandler
     public void on(PermissionCreatedEvent permissionCreatedEvent) {
         Permission permission = permissionMapper.toEntity(permissionCreatedEvent);
         permissionRepository.save(permission);
     }
 
+    @Transactional
     @EventHandler
     public void on(PermissionUpdatedEvent permissionUpdatedEvent) {
         Permission permission = permissionRepository.findById(permissionUpdatedEvent.getPermissionId())
@@ -34,8 +38,12 @@ public class PermissionEventHandler {
         permissionRepository.save(permission);
     }
 
+    @Transactional
     @EventHandler
     public void on(PermissionDeletedEvent permissionDeletedEvent) {
+        if(permissionRepository.isPermissionInUse(permissionDeletedEvent.getPermissionId())) {
+            throw new ResourceInUseException(ErrorCode.PERMISSION_IN_USE);
+        }
         Permission permission = permissionRepository.findById(permissionDeletedEvent.getPermissionId())
                 .orElseThrow(() -> new ResourceNotFoundException(ErrorCode.PERMISSION_NOT_FOUND));
         permissionRepository.delete(permission);
